backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Data & ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# AI Integration
groq==0.4.1

# Testing
pytest==7.4.4
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
